Moving Python command-line argument helper function to contrib folder.
